# alodo-backend/public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "ALODO — API du diagnostic MPME",
        "description": "Documentation générée depuis les annotations des contrôleurs. Obtenir data.token via login/register, puis utiliser Authorize. Les identifiants des exemples sont fictifs.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Backend courant"
        }
    ],
    "paths": {
...
        "/api/store/domains": {
            "post": {
                "tags": [
                    "Administration"
                ],
                "summary": "Créer un domaine",
                "description": "QuestionsController::storeDomain. Rôle admin uniquement. ordre doit être unique entre 1 et 8 ; un intitulé déjà utilisé produit 400. Général doit rester is_scored=false conformément au barème. Aucune route de liste, modification ou suppression administrateur.",
                "operationId": "storeDomain",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DomainRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Domain"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "422": {
                        "$ref": "#/components/responses/Validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/store/questions": {
            "post": {
                "tags": [
                    "Administration"
                ],
                "summary": "Créer une question",
                "description": "QuestionsController::storeQuestion. Rôle admin uniquement. Options requises pour les choix : value et label non vides, value distinctes. Ne pas envoyer d’options pour text/number. question_code n’est pas accepté par la validation actuelle. Ajouter une question sans adapter le code et le barème peut empêcher toute finalisation ; cet endpoint ne configure pas le scoring.",
                "operationId": "storeQuestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuestionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "success",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "enum": [
                                                true
                                            ]
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Question"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/Validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Error": {
                "description": "Les erreurs Laravel peuvent ne contenir que message. Les contrôleurs peuvent ajouter success:false et errors:null.",
                "required": [
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            false
                        ]
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "enum": [
                            false
                        ]
                    },
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object",
                "example": {
                    "message": "The given data was invalid.",
                    "errors": {
                        "responses.0.valeur": [
                            "The selected responses.0.valeur is invalid."
                        ]
                    }
                }
            },
            "Option": {
                "required": [
                    "value",
                    "label"
                ],
                "properties": {
                    "value": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1
                    },
                    "label": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "DomainRequest": {
                "required": [
                    "intitule",
                    "ordre",
                    "is_scored"
                ],
                "properties": {
                    "intitule": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "ordre": {
                        "description": "Unique parmi tous les domaines.",
                        "type": "integer",
                        "maximum": 8,
                        "minimum": 1
                    },
                    "is_scored": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Domain": {
                "required": [
                    "id",
                    "intitule",
                    "ordre",
                    "is_scored"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "intitule": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 1
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "ordre": {
                        "description": "Unique parmi tous les domaines.",
                        "type": "integer",
                        "maximum": 8,
                        "minimum": 1
                    },
                    "is_scored": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "QuestionRequest": {
                "description": "Options obligatoires pour les choix ; omettre options pour texte/nombre. question_code ne fait pas partie des champs validés par cet endpoint.",
                "oneOf": [
                    {
                        "required": [
                            "intitule",
                            "domain_id",
                            "ordre",
                            "type",
                            "options"
                        ],
                        "properties": {
                            "intitule": {
                                "type": "string",
                                "maxLength": 255,
                                "minLength": 1
                            },
                            "domain_id": {
                                "type": "integer",
                                "minimum": 1,
                                "example": 1
                            },
                            "ordre": {
                                "type": "integer",
                                "minimum": 1
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "unique_choice",
                                    "multiple_choice"
                                ]
                            },
                            "options": {
                                "description": "Les value doivent être distinctes dans une même question.",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Option"
                                },
                                "minItems": 2
                            }
                        },
                        "type": "object"
                    },
                    {
                        "required": [
                            "intitule",
                            "domain_id",
                            "ordre",
                            "type"
                        ],
                        "properties": {
                            "intitule": {
                                "type": "string",
                                "maxLength": 255,
                                "minLength": 1
                            },
                            "domain_id": {
                                "type": "integer",
                                "minimum": 1,
                                "example": 1
                            },
                            "ordre": {
                                "type": "integer",
                                "minimum": 1
                            },
                            "type": {
                                "type": "string",
                                "enum": [
                                    "text",
                                    "number"
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Question": {
                "required": [
                    "id",
                    "intitule",
                    "domain_id",
                    "ordre",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "intitule": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1
                    },
                    "domain_id": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "ordre": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "unique_choice",
                            "multiple_choice",
                            "text",
                            "number"
                        ]
                    },
                    "options": {
                        "description": "Les value doivent être distinctes dans une même question.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        },
                        "minItems": 2,
                        "nullable": true
                    },
                    "question_code": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "401 : token absent/invalide ou identifiants de connexion incorrects.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "403 : rôle non autorisé.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "404 : ressource absente, diagnostic inaccessible ou rôle non configuré.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Validation": {
                "description": "422 : données invalides. Les clés responses.{index}.valeur concernent la soumission ; answers.{question_id} concerne la complétude finale. diagnostic_id peut signaler un diagnostic déjà terminé.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "BadRequest": {
                "description": "400 : un domaine de même intitulé existe déjà.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Laravel Sanctum, sans préfixe Bearer dans Authorize.",
                "scheme": "bearer"
            }
        }
    },
...
}